Added level data, reworked end of level 2
I added a new data collection system for the different levels. First I added labels for the custom events which change depending on the level.  I also added a bool which can be changed in the inspector to tell if it is a tester playing or a dev.  I also reworked the end of level 2 as I felt it killed pace within the game.